Change some ints and chars to unsigned
Convert some of the int and char variables to be unsigned int and unsigned
char types.  This prevents gcc generating the warnings "comparison between
signed and unsigned integer expressions" and "array subscript has type
'char'", as well as having a nice side-effect of shaving 42 bytes off the
compiled size when building for the Uno.